Fix typeset -u/-l on NetBSD
On NetBSD, for some reason, the wctrans(3) and towctrans(3) C
library functions exist, but have no effect; the "toupper" and
"tolower" maps don't even translate case for ASCII, never mind wide
characters. This kills 'typeset -u' and 'typeset -l' on ksh, which
was the cause of most of the regression test failures on NetBSD.
Fallback versions for these functions are provided in init.c, but
were not being used on NetBSD because the feature test detected the
presence of these functions in the C library.

src/cmd/ksh93/features/locale:
- Replace the simple test for the presence of wctrans(3),
  towctrans(3), and the wctrans_t type by an actual feature test
  that checks that these functions not only compile, but are also
  capable of changing an ASCII 'q' to upper case and back again.

src/cmd/ksh93/sh/init.c: towctrans():
- Add wide character support to the fallback function, for whatever
  good that may do; on NetBSD, the wide-character towupper(3) and
  towlower(3) functions only change case for ASCII.