tests/types.sh: revert warning (re: e311f1e6)
That regression, which was consistent on the github CI runners, has
not been back in the last nine commits' CI runs, even with all the
new regex code active. I've no idea what fixed it. Perhaps c5018f7c
had something to do with it, as that's the only recent change with
any relevance to 'typeset -T'. In any case, life is too short to
track down intermittent failures that disappeared. Let's just turn
a test failure back into an actual failure so we don't fail to
notice if the regression ever shows up again.

Closes: https://github.com/ksh93/ksh/issues/400 (at least for now)