Restore namespace's parent scope when exiting due to error
Reproducer:

    $ namespace test { x=123; typeset -g x=456; }
    $ echo $x ${.test.x}
    456 123
    $ namespace test { typeset -Q; }
    arch/darwin.i386-64/bin/ksh: typeset: -Q: unknown option
    [usage message snipped for brevity]
    $ echo $x ${.test.x}
    123 123	            <== expected: 123 456
    $ x=789
    $ echo $x ${.test.x}
    789 789                 <== expected: 789 456
    $ # look at that, we never left the namespace...

When prefixing the erroneous 'typeset' with 'command', the problem
does not occur. 'command' disables the properties of special
built-ins such as exit on error. So, when a special built-in exits
on error, the parent scope is not properly resotred.

This bug exists in every ksh93 version with SHOPT_NAMESPACE so far.

src/cmd/ksh93/sh/xec.c: sh_exec():
- Before entering a namespace, use sh_pushcontext and sigsetjmp to
  make sure we return here if sh_exit() is called, e.g. when a
  special builtin throws an error, to ensure the parent scope
  (oldnspace) is restored.

Thanks to @hyenias for making me aware of this bug.
Discussion: https://github.com/ksh93/ksh/issues/479#issuecomment-1140468965